projects_access refactor
